Moving parse and response mapping to factory
